refactor: extract some types from index.d.ts to a more appropriate location
